refactor: split bot.api into bot_instance, error_reporter, and lifecycle modules
Move get_bot/set_bot_instance into bot.core.bot_instance and send_error_to_discord
into bot.core.error_reporter so services and cogs only import what they need,
decoupling them from the full bot lifecycle module.
