+++ backend/bot/cogs/ride_coverage.py
@@ -5,8 +5,8 @@ import logging
 import discord
 from discord.ext import commands
 
-from bot.api import send_error_to_discord
 from bot.core.enums import ChannelIds
+from bot.core.error_reporter import send_error_to_discord
 from bot.repositories.ride_coverage_repository import RideCoverageRepository
 from bot.utils.time_helpers import get_last_sunday
